move check_labels function
